Add search form on services result page
